backup/dfarm-v2/requirements.txt
pyserial
wiringpi
raspberrypi-tm1637
adafruit-circuitpython-am2320